Added some test cases for logs that check whether the logs are received successfully using the turn manager and the worker.
