+++ aimmo-game/simulation/turn_manager.py
@@ -57,14 +57,10 @@ class TurnManager(Thread):
         Send an avatar its view of the game state and register its
         chosen action & logs.
         """
-        print "INSIDE REGISTER ACTION"
         with state_provider as game_state:
             state_view = game_state.get_state_for(avatar)
 
         worker_data = avatar.fetch_data(state_view)
-
-        print worker_data
-
         if avatar.decide_action(worker_data):
             with state_provider as game_state:
                 avatar.action.register(game_state.world_map)
